Refactored Campaign Manager hooks, operators, and sensors to match product naming conventions.
